darija/backend/requirements.txt
# FastAPI & Lambda
fastapi>=0.104.0
mangum>=0.17.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Validation & Config
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.1.0

# HTTP Client (Claude API)
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
aiosqlite>=0.19.0
